[util] Move `Timer` to its own `utils.timer` module
`util.py` is deprecated, and `Timer` was still used in a check
(vsphere). Move to `utils`, and still import it in `util` so that
custom checks importing it from `util` still work (until we finally
remove it in Agent 6.0)
