+++ checks/collector.py
@@ -34,15 +34,13 @@ from config import get_system_stats, get_version
 import checks.system.unix as u
 import checks.system.win32 as w32
 import modules
-from util import (
-    get_uuid,
-    Timer,
-)
+from util import get_uuid
 from utils.cloud_metadata import GCE, EC2
 from utils.logger import log_exceptions
 from utils.jmx import JMXFiles
 from utils.platform import Platform, get_os
 from utils.subprocess_output import get_subprocess_output
+from utils.timer import Timer
 
 log = logging.getLogger(__name__)
